clippy を0件にする
CIでゲートにする前の掃除。機械的な --fix に加えて:

- kbd_shift / kbd_peeked を削除 — 書くだけで誰も読まない死骸だった
  (キーボードの状態をBDAへ移した名残)。毎ビルド警告が出ていた
- 意図的な書き方には理由つきで allow を付けた:
  - tests/interrupts.rs の `1 * 4` は「ベクタ番号 × 表の1要素4バイト」。
    --fix が `4` に畳んでしまったので**戻して** allow (畳むと読めない)
  - bios.rs scroll_window の引数8個は INT 10h のレジスタ割り当てそのまま

94テスト通過。

+++ core/tests/bios.rs
@@ -346,7 +346,7 @@ fn int10_writes_a_string() {
         m.write8(SRC + i as u32, *c);
     }
     m.cpu.sregs[0] = 0; // ES
-    m.cpu.regs[5] = SRC as u32; // BP
+    m.cpu.regs[5] = SRC; // BP
     m.cpu.regs[AX] = 0x1300;
     m.cpu.regs[BX] = 0x0A00; // 明るい緑
     m.cpu.regs[CX] = 2;
